requirements.txt
# Database connectors
mysql-connector-python>=8.2.0
psycopg2-binary>=2.9.9
pymongo>=4.6.0

# Cloud storage
boto3>=1.34.0
google-cloud-storage>=2.14.0
azure-storage-blob>=12.19.0

# Notifications
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0

